pyproject.toml
[project]
name = "anp-open-sdk"
version = "0.1.0"
description = ""
authors = [
    {name = "seanzhang9999",email = "seanzhang9999@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyyaml (>=6.0.2,<7.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pycryptodome (>=3.23.0,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "cryptography (>=43.0.3,<44.0.0)",
    "fastapi (>=0.115.13,<0.116.0)",
    "aiohttp (>=3.12.13,<4.0.0)",
    "base58 (>=2.1.1,<3.0.0)",
    "nest-asyncio (>=1.6.0,<2.0.0)",
    "openai (>=1.91.0,<2.0.0)",
    "uvicorn[standard] (>=0.34.3,<0.35.0)",
    "colorama (>=0.4.6,<0.5.0)",
    "aiofiles (>=24.1.0,<25.0.0)",
    "jcs (>=0.2.1,<0.3.0)",
    "agent-connect (>=0.3.5,<0.4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
